+++ src/main/scala/com/microsoft/hyperspace/index/dataskipping/execution/DataSkippingFileIndex.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.microsoft.hyperspace.index.dataskipping.util
+package com.microsoft.hyperspace.index.dataskipping.execution
 
 import org.apache.hadoop.fs.Path
 import org.apache.spark.sql.{Column, DataFrame, SparkSession}
@@ -25,6 +25,10 @@ import org.apache.spark.sql.types.StructType
 
 import com.microsoft.hyperspace.index.{FileIdTracker, IndexConstants}
 
+/**
+ * FileIndex implementation that filters out files using the index data and
+ * index predicates generated by applying a data skipping index to the source.
+ */
 class DataSkippingFileIndex(
     sparkSession: SparkSession,
     fileIdTracker: FileIdTracker,
